+++ src/components/course/CourseCard.tsx
