+++ src/components/course/ModelPreview.tsx
