+++ src/pages/ExplorePage.tsx
@@ -1,6 +1,6 @@
 import Footer from '@/components/layout/Footer';
 import ExploreSidebar from '@/components/explore/ExploreSidebar';
-import CourseCard from '@/components/explore/CourseCard';
+import CourseCard from '@/components/course/CourseCard';
 import { Search } from 'lucide-react';
 import { useState } from 'react';
 import { TAGS } from '@/constants/explore';
